rename shipper package to boltdb-shipper for clarity
