+++ pkg/logcli/query/query_test.go
@@ -23,8 +23,8 @@ import (
 	"github.com/grafana/loki/pkg/storage"
 	"github.com/grafana/loki/pkg/storage/chunk/client/local"
 	"github.com/grafana/loki/pkg/storage/config"
+	"github.com/grafana/loki/pkg/storage/stores/boltdbshipper"
 	"github.com/grafana/loki/pkg/storage/stores/indexshipper"
-	"github.com/grafana/loki/pkg/storage/stores/shipper"
 	"github.com/grafana/loki/pkg/util/marshal"
 )
 
@@ -601,7 +601,7 @@ func TestLoadFromURL(t *testing.T) {
 	require.Error(t, err)
 	require.Nil(t, client)
 
-	conf.StorageConfig.BoltDBShipperConfig = shipper.Config{
+	conf.StorageConfig.BoltDBShipperConfig = boltdbshipper.Config{
 		Config: indexshipper.Config{
 			SharedStoreType: config.StorageTypeFileSystem,
 		},
