+++ pkg/querier/queryrange/queryrange.proto
@@ -7,7 +7,7 @@ import "google/protobuf/timestamp.proto";
 import "pkg/logproto/logproto.proto";
 import "pkg/logqlmodel/stats/stats.proto";
 import "pkg/querier/queryrange/queryrangebase/queryrange.proto";
-import "pkg/storage/stores/shipper/indexgateway/indexgatewaypb/gateway.proto";
+import "pkg/storage/stores/boltdbshipper/indexgateway/indexgatewaypb/gateway.proto";
 
 option (gogoproto.marshaler_all) = true;
 option (gogoproto.unmarshaler_all) = true;
@@ -136,7 +136,7 @@ message LokiPromResponse {
 }
 
 message IndexStatsResponse {
-  indexgatewaypb.IndexStatsResponse response = 1 [(gogoproto.customtype) = "github.com/grafana/loki/pkg/storage/stores/shipper/indexgateway/indexgatewaypb.IndexStatsResponse"];
+  indexgatewaypb.IndexStatsResponse response = 1 [(gogoproto.customtype) = "github.com/grafana/loki/pkg/storage/stores/boltdbshipper/indexgateway/indexgatewaypb.IndexStatsResponse"];
   repeated queryrangebase.PrometheusResponseHeader Headers = 2 [
     (gogoproto.jsontag) = "-",
     (gogoproto.customtype) = "github.com/grafana/loki/pkg/querier/queryrange/queryrangebase.PrometheusResponseHeader"
