+++ pkg/storage/stores/boltdbshipper/index/compactor/compacted_index_test.go
@@ -16,8 +16,8 @@ import (
 	"github.com/grafana/loki/pkg/storage/chunk"
 	"github.com/grafana/loki/pkg/storage/chunk/client/local"
 	"github.com/grafana/loki/pkg/storage/config"
+	"github.com/grafana/loki/pkg/storage/stores/boltdbshipper/index/indexfile"
 	"github.com/grafana/loki/pkg/storage/stores/indexshipper/compactor/retention"
-	"github.com/grafana/loki/pkg/storage/stores/shipper/index/indexfile"
 	util_log "github.com/grafana/loki/pkg/util/log"
 )
