+++ pkg/storage/stores/boltdbshipper/index/compactor/util_test.go
@@ -22,9 +22,9 @@ import (
 	"github.com/grafana/loki/pkg/storage/chunk/client/local"
 	chunk_util "github.com/grafana/loki/pkg/storage/chunk/client/util"
 	"github.com/grafana/loki/pkg/storage/config"
+	"github.com/grafana/loki/pkg/storage/stores/boltdbshipper"
+	shipper_util "github.com/grafana/loki/pkg/storage/stores/boltdbshipper/util"
 	"github.com/grafana/loki/pkg/storage/stores/indexshipper"
-	"github.com/grafana/loki/pkg/storage/stores/shipper"
-	shipper_util "github.com/grafana/loki/pkg/storage/stores/shipper/util"
 	util_log "github.com/grafana/loki/pkg/util/log"
 	"github.com/grafana/loki/pkg/validation"
 )
@@ -233,7 +233,7 @@ func newTestStore(t testing.TB, clientMetrics storage.ClientMetrics) *testStore 
 		},
 		MaxParallelGetChunk: 150,
 
-		BoltDBShipperConfig: shipper.Config{
+		BoltDBShipperConfig: boltdbshipper.Config{
 			Config: indexshipper.Config{
 				ActiveIndexDirectory: indexDir,
 				SharedStoreType:      "filesystem",
