+++ pkg/storage/stores/boltdbshipper/indexgateway/indexgatewaypb/gateway.proto
@@ -5,7 +5,7 @@ package indexgatewaypb;
 import "github.com/gogo/protobuf/gogoproto/gogo.proto";
 import "pkg/logproto/logproto.proto";
 
-option go_package = "github.com/grafana/loki/pkg/storage/stores/shipper/indexgateway/indexgatewaypb";
+option go_package = "github.com/grafana/loki/pkg/storage/stores/boltdbshipper/indexgateway/indexgatewaypb";
 
 service IndexGateway {
   /// QueryIndex reads the indexes required for given query & sends back the batch of rows
