+++ pkg/storage/stores/boltdbshipper/shipper_index_client.go
@@ -1,4 +1,4 @@
-package shipper
+package boltdbshipper
 
 import (
 	"context"
@@ -19,11 +19,11 @@ import (
 	"github.com/grafana/loki/pkg/storage/chunk/client/local"
 	"github.com/grafana/loki/pkg/storage/chunk/client/util"
 	"github.com/grafana/loki/pkg/storage/config"
+	"github.com/grafana/loki/pkg/storage/stores/boltdbshipper/index"
+	"github.com/grafana/loki/pkg/storage/stores/boltdbshipper/index/indexfile"
 	"github.com/grafana/loki/pkg/storage/stores/indexshipper"
 	"github.com/grafana/loki/pkg/storage/stores/indexshipper/downloads"
 	series_index "github.com/grafana/loki/pkg/storage/stores/series/index"
-	"github.com/grafana/loki/pkg/storage/stores/shipper/index"
-	"github.com/grafana/loki/pkg/storage/stores/shipper/index/indexfile"
 	util_log "github.com/grafana/loki/pkg/util/log"
 )
