+++ pkg/storage/stores/indexshipper/compactor/deletion/delete_requests_table_test.go
@@ -9,10 +9,10 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/grafana/loki/pkg/storage/chunk/client/local"
+	"github.com/grafana/loki/pkg/storage/stores/boltdbshipper/testutil"
+	"github.com/grafana/loki/pkg/storage/stores/boltdbshipper/util"
 	"github.com/grafana/loki/pkg/storage/stores/indexshipper/storage"
 	"github.com/grafana/loki/pkg/storage/stores/series/index"
-	"github.com/grafana/loki/pkg/storage/stores/shipper/testutil"
-	"github.com/grafana/loki/pkg/storage/stores/shipper/util"
 )
 
 func TestDeleteRequestsTable(t *testing.T) {
